SMALL refactor
CREATE a folder for helper function
