+++ server/tools/helperFunctions.js
